day16 - part 2, clean up

+++ src/main/kotlin/template/Template.kt
@@ -16,10 +16,10 @@ fun main() {
 
 }
 
-fun part1(input: String): Int {
-    return 0
+fun part1(input: String): Long {
+    return 0L
 }
 
-fun part2(input: String): Int {
-    return 0
+fun part2(input: String): Long {
+    return 0L
 }
